Recuperación part1

+++ src/controllers/products.controller.js
@@ -8,19 +8,18 @@ export const findProductById = async (req, res) => {
     try {
         const product = await productService.findById(pid);
         if (!product) {
-            //return res.status(404).json({ message: "Product not found with the id provided" });
-            return CustomError.generateError(errorsMessages.PRODUCT_NOT_FOUND,404)
+            return CustomError.generateError(errorsMessages.PRODUCT_NOT_FOUND, 404)
         }
         res.status(200).json({ message: "Product found", product });
-        } catch (error) {
-            res.status(500).json({ message: error.message });
-        }
+    } catch (error) {
+        res.status(500).json({ message: error.message });
+    }
 };
 
 export const findAllProduct = async (req, res) => {
     try {
-        const products = await productService.findAll(req.query); 
-        res.status(200).json({ message: "Product found", products });   
+        const products = await productService.findAll(req.query);
+        res.status(200).json({ message: "Product found", products });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -28,12 +27,21 @@ export const findAllProduct = async (req, res) => {
 
 export const createOneProduc = async (req, res) => {
     const { title, description, price, code, stock, category } = req.body;
-    if (!title || !description || !price || !code || !stock || ! category) {
-        return res.status(400).json({ message: "Some data is missing" });
+    if (!title || !description || !price || !code || !stock || !category) {
+        return CustomError.generateError(errorsMessages.ALL_FIELDS, 404)
     }
     try {
-        const response = await productService.createOne(req.body);
-        res.status(200).json({ message: "Producto created", response });
+        if (req.user.role === "PREMIUM") {
+            const producto = { ...req.body, owner: req.user.email };
+            const response = await productService.createOne(producto);
+            res.status(200).json({ message: "Producto creado", response });
+
+        } else {
+            const response = await productService.createOne(req.body);
+            res.status(200).json({ message: "Producto creado", response });
+
+        }
+
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -41,16 +49,22 @@ export const createOneProduc = async (req, res) => {
 
 export const deleteOneProdAll = async (req, res) => {
     const { pid } = req.params;
+    const pDelet = await productService.findById(pid)
     try {
-        const response = await productService.deleteOneProduct(pid);
-        if (!response) {
-            //return res.status(404).json({ message: "Product not found with the id provided" });
-            return CustomError.generateError(errorsMessages.PRODUCT_NOT_FOUND,404)
+        if (req.user.role === "PREMIUM") {
+            if (pDelet.owner === req.user._id) {
+                const response = await productService.deleteOneProduct(pid);
+                if (!response) {
+                    return CustomError.generateError(errorsMessages.PRODUCT_NOT_FOUND, 404)
+                }
+                res.status(200).json({ message: "Product deleted" });
+            }
+        } else {
+            return res.status(500).json({message: "Producto creado por otro usuario."})
         }
-        res.status(200).json({ message: "Product deleted" });
-        } catch (error) {
+    } catch (error) {
         res.status(500).json({ message: error.message });
-        }
+    }
 }
 
 export const updateProducts = async (req, res) => {
@@ -58,7 +72,7 @@ export const updateProducts = async (req, res) => {
     try {
         const response = await productService.updateProduct(pid, req.body);
         if (!response) {
-            return CustomError.generateError(errorsMessages.PRODUCT_NOT_FOUND,404)
+            return CustomError.generateError(errorsMessages.PRODUCT_NOT_FOUND, 404)
         }
         res.status(200).json({ message: "Product updated", response });
     } catch (error) {
@@ -69,9 +83,9 @@ export const updateProducts = async (req, res) => {
 export const productMocksController = async (req, res, next) => {
     try {
         const mockData = productMock();
-  
+
         res.status(200).json({ message: "Product created successfully", mockData });
     } catch (error) {
-      res.status(500).json({ message: error.message });
+        res.status(500).json({ message: error.message });
     }
-  }
+}
